+++ src/frontend/apps/impress/src/features/docs/docs-grid/components/DocsGrid.tsx
@@ -1,14 +1,21 @@
-import { Button } from '@gouvfr-lasuite/cunningham-react';
-import { useMemo } from 'react';
+import {
+  Button,
+  VariantType,
+  useToastProvider,
+} from '@gouvfr-lasuite/cunningham-react';
+import { useMemo, useState } from 'react';
+import { useDropzone } from 'react-dropzone';
 import { useTranslation } from 'react-i18next';
 import { InView } from 'react-intersection-observer';
 import { css } from 'styled-components';
 
-import { Box, Card, Text } from '@/components';
+import AllDocs from '@/assets/icons/doc-all.svg';
+import { Box, Card, Icon, Text } from '@/components';
 import { DocDefaultFilter, useInfiniteDocs } from '@/docs/doc-management';
 import { useResponsiveStore } from '@/stores';
 
 import { useInfiniteDocsTrashbin } from '../api';
+import { ContentTypesAllowed, useImportDoc } from '../api/useImportDoc';
 import { useResponsiveDocGrid } from '../hooks/useResponsiveDocGrid';
 
 import {
@@ -24,6 +31,70 @@ export const DocsGrid = ({
   target = DocDefaultFilter.ALL_DOCS,
 }: DocsGridProps) => {
   const { t } = useTranslation();
+  const [isDragOver, setIsDragOver] = useState(false);
+  const { toast } = useToastProvider();
+
+  const MAX_FILE_SIZE = {
+    bytes: 10 * 1024 * 1024, // 10 MB
+    text: '10MB',
+  };
+
+  const { getRootProps, getInputProps, open } = useDropzone({
+    accept: {
+      [ContentTypesAllowed.Docx]: ['.docx'],
+      [ContentTypesAllowed.Markdown]: ['.md'],
+    },
+    maxSize: MAX_FILE_SIZE.bytes,
+    onDrop(acceptedFiles) {
+      setIsDragOver(false);
+      for (const file of acceptedFiles) {
+        importDoc(file);
+      }
+    },
+    onDragEnter: () => {
+      setIsDragOver(true);
+    },
+    onDragLeave: () => {
+      setIsDragOver(false);
+    },
+    onDropRejected(fileRejections) {
+      fileRejections.forEach((rejection) => {
+        const isFileTooLarge = rejection.errors.some(
+          (error) => error.code === 'file-too-large',
+        );
+
+        if (isFileTooLarge) {
+          toast(
+            t(
+              'The document "{{documentName}}" is too large. Maximum file size is {{maxFileSize}}.',
+              {
+                documentName: rejection.file.name,
+                maxFileSize: MAX_FILE_SIZE.text,
+              },
+            ),
+            VariantType.ERROR,
+          );
+        } else {
+          toast(
+            t(
+              `The document "{{documentName}}" import has failed (only .docx and .md files are allowed)`,
+              {
+                documentName: rejection.file.name,
+              },
+            ),
+            VariantType.ERROR,
+          );
+        }
+      });
+    },
+    noClick: true,
+  });
+  const { mutate: importDoc } = useImportDoc();
+
+  const withUpload =
+    !target ||
+    target === DocDefaultFilter.ALL_DOCS ||
+    target === DocDefaultFilter.MY_DOCS;
 
   const { isDesktop } = useResponsiveStore();
   const { flexLeft, flexRight } = useResponsiveDocGrid();
@@ -60,21 +131,6 @@ export const DocsGrid = ({
     void fetchNextPage();
   };
 
-  let title = t('All docs');
-  switch (target) {
-    case DocDefaultFilter.MY_DOCS:
-      title = t('My docs');
-      break;
-    case DocDefaultFilter.SHARED_WITH_ME:
-      title = t('Shared with me');
-      break;
-    case DocDefaultFilter.TRASHBIN:
-      title = t('Trashbin');
-      break;
-    default:
-      title = t('All docs');
-  }
-
   return (
     <Box
       $position="relative"
@@ -91,16 +147,24 @@ export const DocsGrid = ({
         $width="100%"
         $css={css`
           ${!isDesktop ? 'border: none;' : ''}
+          ${isDragOver
+            ? `
+              border: 2px dashed var(--c--contextuals--border--semantic--brand--primary);
+              background-color: var(--c--contextuals--background--semantic--brand--tertiary);
+            `
+            : ''}
         `}
         $padding={{
-          top: 'base',
-          horizontal: isDesktop ? 'md' : 'xs',
           bottom: 'md',
         }}
+        {...(withUpload ? getRootProps({ className: 'dropzone' }) : {})}
       >
-        <Text as="h2" $size="h4" $margin={{ top: '0px', bottom: '10px' }}>
-          {title}
-        </Text>
+        {withUpload && <input {...getInputProps()} />}
+        <DocGridTitleBar
+          target={target}
+          onUploadClick={open}
+          withUpload={withUpload}
+        />
 
         {!hasDocs && !loading && (
           <Box $padding={{ vertical: 'sm' }} $align="center" $justify="center">
@@ -110,7 +174,11 @@ export const DocsGrid = ({
           </Box>
         )}
         {hasDocs && (
-          <Box $gap="6px" $overflow="auto">
+          <Box
+            $gap="6px"
+            $overflow="auto"
+            $padding={{ vertical: 'sm', horizontal: isDesktop ? 'md' : 'xs' }}
+          >
             <Box role="grid" aria-label={t('Documents grid')}>
               <Box role="rowgroup">
                 <Box
@@ -168,6 +236,67 @@ export const DocsGrid = ({
           </Box>
         )}
       </Card>
+    </Box>
+  );
+};
+
+const DocGridTitleBar = ({
+  target,
+  onUploadClick,
+  withUpload,
+}: {
+  target: DocDefaultFilter;
+  onUploadClick: () => void;
+  withUpload: boolean;
+}) => {
+  const { t } = useTranslation();
+  const { isDesktop } = useResponsiveStore();
+
+  let title = t('All docs');
+  let icon = <Icon icon={<AllDocs width={24} height={24} />} />;
+  if (target === DocDefaultFilter.MY_DOCS) {
+    icon = <Icon iconName="lock" />;
+    title = t('My docs');
+  } else if (target === DocDefaultFilter.SHARED_WITH_ME) {
+    icon = <Icon iconName="group" />;
+    title = t('Shared with me');
+  } else if (target === DocDefaultFilter.TRASHBIN) {
+    icon = <Icon iconName="delete" />;
+    title = t('Trashbin');
+  }
+
+  return (
+    <Box
+      $direction="row"
+      $padding={{
+        vertical: 'md',
+        horizontal: isDesktop ? 'md' : 'xs',
+      }}
+      $css={css`
+        border-bottom: 1px solid var(--c--contextuals--border--surface--primary);
+      `}
+      $align="center"
+      $justify="space-between"
+    >
+      <Box $direction="row" $gap="xs" $align="center">
+        {icon}
+        <Text as="h2" $size="h4" $margin="none">
+          {title}
+        </Text>
+      </Box>
+      {withUpload && (
+        <Button
+          color="brand"
+          variant="tertiary"
+          onClick={(e) => {
+            e.stopPropagation();
+            onUploadClick();
+          }}
+          aria-label={t('Open the upload dialog')}
+        >
+          <Icon iconName="upload_file" $withThemeInherited />
+        </Button>
+      )}
     </Box>
   );
 };